Handle ST commands in ipc module
