+++ mullvad-daemon/src/ipc/mod.rs
@@ -4,13 +4,21 @@ use futures::channel::oneshot;
 
 mod api;
 mod geoip;
+mod split_tunnel;
 mod tunnel_state;
 
 // TODO: Move all of `Daemon::handle_command` here.
 
 impl<L: EventListener + Send + Clone + 'static> Daemon<L> {
-    pub(super) async fn handle_ipc_command(&mut self, command: DaemonCommand) {
+    pub(super) async fn handle_ipc_command(&mut self, mut command: DaemonCommand) {
         use DaemonCommand::*;
+
+        #[cfg(any(target_os = "linux", target_os = "windows"))]
+        match self.handle_split_tunnel_command(command).await {
+            CommandResult::Handled => return,
+            CommandResult::NotHandled(cmd) => command = cmd,
+        }
+
         match command {
             GetCurrentLocation(tx) => self.on_get_current_location(tx),
 
@@ -23,6 +31,11 @@ impl<L: EventListener + Send + Clone + 'static> Daemon<L> {
             _ => unreachable!("handled by Daemon::handle_command"),
         }
     }
+}
+
+enum CommandResult {
+    NotHandled(DaemonCommand),
+    Handled,
 }
 
 fn oneshot_send<T>(tx: oneshot::Sender<T>, t: T, msg: &'static str) {
