+++ mullvad-daemon/src/ipc/split_tunnel/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(target_os = "linux")]
+#[path = "linux.rs"]
+pub(super) mod imp;
+
+#[cfg(windows)]
+#[path = "windows.rs"]
+pub(super) mod imp;
