+++ mullvad-daemon/src/lib.rs
@@ -1213,22 +1213,6 @@ where
             GetCurrentVersion(tx) => self.on_get_current_version(tx),
             #[cfg(not(target_os = "android"))]
             FactoryReset(tx) => self.on_factory_reset(tx).await,
-            #[cfg(target_os = "linux")]
-            GetSplitTunnelProcesses(tx) => self.on_get_split_tunnel_processes(tx),
-            #[cfg(target_os = "linux")]
-            AddSplitTunnelProcess(tx, pid) => self.on_add_split_tunnel_process(tx, pid),
-            #[cfg(target_os = "linux")]
-            RemoveSplitTunnelProcess(tx, pid) => self.on_remove_split_tunnel_process(tx, pid),
-            #[cfg(target_os = "linux")]
-            ClearSplitTunnelProcesses(tx) => self.on_clear_split_tunnel_processes(tx),
-            #[cfg(windows)]
-            AddSplitTunnelApp(tx, path) => self.on_add_split_tunnel_app(tx, path).await,
-            #[cfg(windows)]
-            RemoveSplitTunnelApp(tx, path) => self.on_remove_split_tunnel_app(tx, path).await,
-            #[cfg(windows)]
-            ClearSplitTunnelApps(tx) => self.on_clear_split_tunnel_apps(tx).await,
-            #[cfg(windows)]
-            SetSplitTunnelState(tx, enabled) => self.on_set_split_tunnel_state(tx, enabled).await,
             #[cfg(target_os = "windows")]
             UseWireGuardNt(tx, state) => self.on_use_wireguard_nt(tx, state).await,
             #[cfg(target_os = "windows")]
@@ -1692,170 +1676,6 @@ where
             }
             Self::oneshot_send(tx, last_error, "factory_reset response");
         }));
-    }
-
-    #[cfg(target_os = "linux")]
-    fn on_get_split_tunnel_processes(&mut self, tx: ResponseTx<Vec<i32>, split_tunnel::Error>) {
-        let result = self.exclude_pids.list().map_err(|error| {
-            log::error!("{}", error.display_chain_with_msg("Unable to obtain PIDs"));
-            error
-        });
-        Self::oneshot_send(tx, result, "get_split_tunnel_processes response");
-    }
-
-    #[cfg(target_os = "linux")]
-    fn on_add_split_tunnel_process(&mut self, tx: ResponseTx<(), split_tunnel::Error>, pid: i32) {
-        let result = self.exclude_pids.add(pid).map_err(|error| {
-            log::error!("{}", error.display_chain_with_msg("Unable to add PID"));
-            error
-        });
-        Self::oneshot_send(tx, result, "add_split_tunnel_process response");
-    }
-
-    #[cfg(target_os = "linux")]
-    fn on_remove_split_tunnel_process(
-        &mut self,
-        tx: ResponseTx<(), split_tunnel::Error>,
-        pid: i32,
-    ) {
-        let result = self.exclude_pids.remove(pid).map_err(|error| {
-            log::error!("{}", error.display_chain_with_msg("Unable to remove PID"));
-            error
-        });
-        Self::oneshot_send(tx, result, "remove_split_tunnel_process response");
-    }
-
-    #[cfg(target_os = "linux")]
-    fn on_clear_split_tunnel_processes(&mut self, tx: ResponseTx<(), split_tunnel::Error>) {
-        let result = self.exclude_pids.clear().map_err(|error| {
-            log::error!("{}", error.display_chain_with_msg("Unable to clear PIDs"));
-            error
-        });
-        Self::oneshot_send(tx, result, "clear_split_tunnel_processes response");
-    }
-
-    /// Update the split app paths in both the settings and tunnel
-    #[cfg(windows)]
-    async fn set_split_tunnel_paths(
-        &mut self,
-        tx: ResponseTx<(), Error>,
-        response_msg: &'static str,
-        settings: Settings,
-        update: ExcludedPathsUpdate,
-    ) {
-        let new_list = match update {
-            ExcludedPathsUpdate::SetPaths(ref paths) => {
-                if *paths == settings.split_tunnel.apps {
-                    Self::oneshot_send(tx, Ok(()), response_msg);
-                    return;
-                }
-                paths.iter()
-            }
-            ExcludedPathsUpdate::SetState(_) => settings.split_tunnel.apps.iter(),
-        };
-        let new_state = match update {
-            ExcludedPathsUpdate::SetPaths(_) => settings.split_tunnel.enable_exclusions,
-            ExcludedPathsUpdate::SetState(state) => {
-                if state == settings.split_tunnel.enable_exclusions {
-                    Self::oneshot_send(tx, Ok(()), response_msg);
-                    return;
-                }
-                state
-            }
-        };
-
-        if new_state || new_state != settings.split_tunnel.enable_exclusions {
-            let tunnel_list = if new_state {
-                new_list.map(|s| OsString::from(s)).collect()
-            } else {
-                vec![]
-            };
-
-            let (result_tx, result_rx) = oneshot::channel();
-            self.send_tunnel_command(TunnelCommand::SetExcludedApps(result_tx, tunnel_list));
-            let daemon_tx = self.tx.clone();
-
-            tokio::spawn(async move {
-                match result_rx.await {
-                    Ok(Ok(_)) => (),
-                    Ok(Err(error)) => {
-                        log::error!(
-                            "{}",
-                            error.display_chain_with_msg("Failed to set excluded apps list")
-                        );
-                        Self::oneshot_send(tx, Err(Error::SplitTunnelError(error)), response_msg);
-                        return;
-                    }
-                    Err(_) => {
-                        log::error!("The tunnel failed to return a result");
-                        return;
-                    }
-                }
-
-                let _ = daemon_tx.send(InternalDaemonEvent::ExcludedPathsEvent(update, tx));
-            });
-        } else {
-            let _ = self
-                .tx
-                .send(InternalDaemonEvent::ExcludedPathsEvent(update, tx));
-        }
-    }
-
-    #[cfg(windows)]
-    async fn on_add_split_tunnel_app(&mut self, tx: ResponseTx<(), Error>, path: PathBuf) {
-        let settings = self.settings.to_settings();
-
-        let mut new_list = settings.split_tunnel.apps.clone();
-        new_list.insert(path);
-
-        self.set_split_tunnel_paths(
-            tx,
-            "add_split_tunnel_app response",
-            settings,
-            ExcludedPathsUpdate::SetPaths(new_list),
-        )
-        .await;
-    }
-
-    #[cfg(windows)]
-    async fn on_remove_split_tunnel_app(&mut self, tx: ResponseTx<(), Error>, path: PathBuf) {
-        let settings = self.settings.to_settings();
-
-        let mut new_list = settings.split_tunnel.apps.clone();
-        new_list.remove(&path);
-
-        self.set_split_tunnel_paths(
-            tx,
-            "remove_split_tunnel_app response",
-            settings,
-            ExcludedPathsUpdate::SetPaths(new_list),
-        )
-        .await;
-    }
-
-    #[cfg(windows)]
-    async fn on_clear_split_tunnel_apps(&mut self, tx: ResponseTx<(), Error>) {
-        let settings = self.settings.to_settings();
-        let new_list = HashSet::new();
-        self.set_split_tunnel_paths(
-            tx,
-            "clear_split_tunnel_apps response",
-            settings,
-            ExcludedPathsUpdate::SetPaths(new_list),
-        )
-        .await;
-    }
-
-    #[cfg(windows)]
-    async fn on_set_split_tunnel_state(&mut self, tx: ResponseTx<(), Error>, state: bool) {
-        let settings = self.settings.to_settings();
-        self.set_split_tunnel_paths(
-            tx,
-            "set_split_tunnel_state response",
-            settings,
-            ExcludedPathsUpdate::SetState(state),
-        )
-        .await;
     }
 
     #[cfg(windows)]
